restructured the file structure
