+++ package.json
@@ -29,6 +29,7 @@
     "globals": "^15.12.0",
     "typescript": "~5.6.2",
     "typescript-eslint": "^8.15.0",
-    "vite": "^6.0.1"
+    "vite": "^6.0.1",
+    "vite-tsconfig-paths": "^5.1.4"
   }
 }
