+++ src/components/EntryForm/Master.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useState } from "react";
-import { useAddStore } from "../../useAddStore";
-import Management from "../Management/Management";
+import { useAddStore } from "@components/EntryList/useAddStore";
+import Management from "@components/EntryList/Management";
 import Form from "../Input/Index";
 
 interface Props {
