+++ src/components/EntryList/MangementTypes.tsx
