+++ src/components/EntryList/Modal.tsx
