+++ src/components/EntryList/types.tsx
