--- src/components/menu/types.ts
@@ -1,6 +1,0 @@
-export interface MenuItem {
-  label: string;
-  children?: MenuItem[];
-  action?: (...args: any[]) => void;
-}
-
