+++ src/main.tsx
@@ -1,10 +1,8 @@
 // import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import './index.css'
-import App from './App.tsx'
+import './styles/index.css'
+import App from '@pages/App'
 
 createRoot(document.getElementById('root')!).render(
-
     <App />
-
 )
