--- src/store/auth/types.ts
@@ -1,5 +1,0 @@
-export type TAuthStore = {
-  token?: string;
-  refreshToken?: string;
-  isLoggedIn?: boolean;
-}
